Fix 500 while deleting non-existing user.

+++ backend/routers/user.py
@@ -18,8 +18,8 @@ def get_user_info(db: Session = Depends(get_db), user_id: str = Depends(oauth2.r
 @router.delete("/{user_id}")
 def delete_user(user_id: uuid.UUID, db: Session = Depends(get_db), current_user: User = Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.id == user_id).first()
-    # if not user:
-    #     raise HTTPException(status_code=404, detail="User not found")
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(user)
     db.commit()
     return {"message": "User deleted"}
